Refactor code to reduce circular dependencies
