+++ lib/spack/spack/util/web.py
@@ -21,23 +21,18 @@ from typing import IO, Optional
 from urllib.error import HTTPError, URLError
 from urllib.request import HTTPSHandler, Request, build_opener
 
-import llnl.util.lang
-import llnl.util.tty as tty
+import llnl.url
+from llnl.util import lang, tty
 from llnl.util.filesystem import mkdirp, rename, working_dir
 
-import spack
 import spack.config
 import spack.error
 import spack.gcs_handler
 import spack.s3_handler
-import spack.url
-import spack.util.crypto
 import spack.util.gcs as gcs_util
 import spack.util.s3 as s3_util
 import spack.util.url as url_util
-from spack.util.compression import ALLOWED_ARCHIVE_TYPES
 from spack.util.executable import CommandNotFoundError, which
-from spack.util.path import convert_to_posix_path
 
 
 class DetailedHTTPError(HTTPError):
@@ -90,7 +85,7 @@ def _urlopen():
 
 
 #: Dispatches to the correct OpenerDirector.open, based on Spack configuration.
-urlopen = llnl.util.lang.Singleton(_urlopen)
+urlopen = lang.Singleton(_urlopen)
 
 #: User-Agent used in Request objects
 SPACK_USER_AGENT = "Spackbot/{0}".format(spack.spack_version)
@@ -636,7 +631,7 @@ def spider(root_urls, depth=0, concurrency=32):
                 links.add(abs_link)
 
                 # Skip stuff that looks like an archive
-                if any(raw_link.endswith(s) for s in ALLOWED_ARCHIVE_TYPES):
+                if any(raw_link.endswith(s) for s in llnl.url.ALLOWED_ARCHIVE_TYPES):
                     continue
 
                 # Skip already-visited links
@@ -696,7 +691,7 @@ def spider(root_urls, depth=0, concurrency=32):
                     current_depth, depth, len(spider_args)
                 )
             )
-            results = tp.map(llnl.util.lang.star(_spider), spider_args)
+            results = tp.map(lang.star(_spider), spider_args)
             spider_args = []
             collect = current_depth < depth
             for sub_pages, sub_links, sub_spider_args in results:
@@ -711,123 +706,6 @@ def spider(root_urls, depth=0, concurrency=32):
         tp.join()
 
     return pages, links
-
-
-def find_versions_of_archive(
-    archive_urls, list_url=None, list_depth=0, concurrency=32, reference_package=None
-):
-    """Scrape web pages for new versions of a tarball. This function prefers URLs in the
-    following order: links found on the scraped page that match a url generated by the
-    reference package, found and in the archive_urls list, found and derived from those
-    in the archive_urls list, and if none are found for a version then the item in the
-    archive_urls list is included for the version.
-
-    Args:
-        archive_urls (str or list or tuple): URL or sequence of URLs for
-            different versions of a package. Typically these are just the
-            tarballs from the package file itself. By default, this searches
-            the parent directories of archives.
-        list_url (str or None): URL for a listing of archives.
-            Spack will scrape these pages for download links that look
-            like the archive URL.
-        list_depth (int): max depth to follow links on list_url pages.
-            Defaults to 0.
-        concurrency (int): maximum number of concurrent requests
-        reference_package (spack.package_base.PackageBase or None): a spack package
-            used as a reference for url detection.  Uses the url_for_version
-            method on the package to produce reference urls which, if found,
-            are preferred.
-    """
-    if not isinstance(archive_urls, (list, tuple)):
-        archive_urls = [archive_urls]
-
-    # Generate a list of list_urls based on archive urls and any
-    # explicitly listed list_url in the package
-    list_urls = set()
-    if list_url is not None:
-        list_urls.add(list_url)
-    for aurl in archive_urls:
-        list_urls |= spack.url.find_list_urls(aurl)
-
-    # Add '/' to the end of the URL. Some web servers require this.
-    additional_list_urls = set()
-    for lurl in list_urls:
-        if not lurl.endswith("/"):
-            additional_list_urls.add(lurl + "/")
-    list_urls |= additional_list_urls
-
-    # Grab some web pages to scrape.
-    pages, links = spider(list_urls, depth=list_depth, concurrency=concurrency)
-
-    # Scrape them for archive URLs
-    regexes = []
-    for aurl in archive_urls:
-        # This creates a regex from the URL with a capture group for
-        # the version part of the URL.  The capture group is converted
-        # to a generic wildcard, so we can use this to extract things
-        # on a page that look like archive URLs.
-        url_regex = spack.url.wildcard_version(aurl)
-
-        # We'll be a bit more liberal and just look for the archive
-        # part, not the full path.
-        # this is a URL so it is a posixpath even on Windows
-        url_regex = PurePosixPath(url_regex).name
-
-        # We need to add a / to the beginning of the regex to prevent
-        # Spack from picking up similarly named packages like:
-        #   https://cran.r-project.org/src/contrib/pls_2.6-0.tar.gz
-        #   https://cran.r-project.org/src/contrib/enpls_5.7.tar.gz
-        #   https://cran.r-project.org/src/contrib/autopls_1.3.tar.gz
-        #   https://cran.r-project.org/src/contrib/matrixpls_1.0.4.tar.gz
-        url_regex = "/" + url_regex
-
-        # We need to add a $ anchor to the end of the regex to prevent
-        # Spack from picking up signature files like:
-        #   .asc
-        #   .md5
-        #   .sha256
-        #   .sig
-        # However, SourceForge downloads still need to end in '/download'.
-        url_regex += r"(\/download)?"
-        # PyPI adds #sha256=... to the end of the URL
-        url_regex += "(#sha256=.*)?"
-        url_regex += "$"
-
-        regexes.append(url_regex)
-
-    # Build a dict version -> URL from any links that match the wildcards.
-    # Walk through archive_url links first.
-    # Any conflicting versions will be overwritten by the list_url links.
-    versions = {}
-    matched = set()
-    for url in sorted(links):
-        url = convert_to_posix_path(url)
-        if any(re.search(r, url) for r in regexes):
-            try:
-                ver = spack.url.parse_version(url)
-                if ver in matched:
-                    continue
-                versions[ver] = url
-                # prevent this version from getting overwritten
-                if reference_package is not None:
-                    if url == reference_package.url_for_version(ver):
-                        matched.add(ver)
-                else:
-                    extrapolated_urls = [
-                        spack.url.substitute_version(u, ver) for u in archive_urls
-                    ]
-                    if url in extrapolated_urls:
-                        matched.add(ver)
-            except spack.url.UndetectableVersionError:
-                continue
-
-    for url in archive_urls:
-        url = convert_to_posix_path(url)
-        ver = spack.url.parse_version(url)
-        if ver not in versions:
-            versions[ver] = url
-
-    return versions
 
 
 def get_header(headers, header_name):
